+++ src/elements/teaser-product/teaser-product/teaser-product.component.ts
@@ -26,7 +26,7 @@ export class SbbTeaserProductElement extends SbbTeaserProductCommonElementMixin(
     return html`
       <div class="sbb-teaser-product__wrapper">
         ${this.renderLink(
-          // For SEO, we add the accessibility hidden as hidden content of the link
+          // For SEO, we add the accessibility label as hidden content of the link
           html`<span class="sbb-screen-reader-only">${this.accessibilityLabel}</span>`,
         )}
         ${this.renderTemplate()}