+++ CHANGELOG.md
@@ -2,6 +2,53 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [5.9.0](https://github.com/sbb-design-systems/lyne-components/compare/v5.8.0...v5.9.0) (2026-09-15)
+
+
+### Features
+
+* **download:** add vertical divider, correct title font-size, support download attribute ([#5236](https://github.com/sbb-design-systems/lyne-components/issues/5236)) ([e0cd632](https://github.com/sbb-design-systems/lyne-components/commit/e0cd632b0b04d213956fa80609b92da64499af76))
+* **logo:** add SBB Cargo and Elvetino logos ([#5240](https://github.com/sbb-design-systems/lyne-components/issues/5240)) ([4e04f38](https://github.com/sbb-design-systems/lyne-components/commit/4e04f383d39b507d7ae79aa08669caa8f55904ab))
+* **seat-reservation:** add db icons ([#5242](https://github.com/sbb-design-systems/lyne-components/issues/5242)) ([fe2161f](https://github.com/sbb-design-systems/lyne-components/commit/fe2161f1a11fd91484a163e39501e05d47c86cf1))
+
+
+### Bug Fixes
+
+* **dialog:** various mobile view optimizations ([#5209](https://github.com/sbb-design-systems/lyne-components/issues/5209)) ([95ce1c5](https://github.com/sbb-design-systems/lyne-components/commit/95ce1c5ebb9c3282130dc7b73c8ca982d148a3a1))
+* **form-field:** increase label space for textarea in 's' size ([#5223](https://github.com/sbb-design-systems/lyne-components/issues/5223)) ([2e3be1f](https://github.com/sbb-design-systems/lyne-components/commit/2e3be1f11f87ad6cb47ae8e3f5857ed363f26104)), closes [#5196](https://github.com/sbb-design-systems/lyne-components/issues/5196)
+* **map-container:** fix keyboard navigation order ([#5245](https://github.com/sbb-design-systems/lyne-components/issues/5245)) ([a0bd958](https://github.com/sbb-design-systems/lyne-components/commit/a0bd9587aca150091ab7d8d9e3f6afbf47955f9c)), closes [#5179](https://github.com/sbb-design-systems/lyne-components/issues/5179)
+
+
+### Styles
+
+* **notification:** improve mobile layout and text alignment ([#5217](https://github.com/sbb-design-systems/lyne-components/issues/5217)) ([fba8dab](https://github.com/sbb-design-systems/lyne-components/commit/fba8dab959227aabe00fbd00116c82491c1d64d9))
+
+## [5.8.0](https://github.com/sbb-design-systems/lyne-components/compare/v5.7.0...v5.8.0) (2026-09-07)
+
+
+### Features
+
+* **popover:** negative variant ([#5213](https://github.com/sbb-design-systems/lyne-components/issues/5213)) ([d305b7a](https://github.com/sbb-design-systems/lyne-components/commit/d305b7abfee70ab3f213c7f952ead740f7f80e52))
+* **select,autocomplete:** expose CSS variable for panel gap ([#5219](https://github.com/sbb-design-systems/lyne-components/issues/5219)) ([ad3ef27](https://github.com/sbb-design-systems/lyne-components/commit/ad3ef270a9264bab60574fa6a65f014e53321438))
+* **teaser:** add new size, alignment and static options ([#5207](https://github.com/sbb-design-systems/lyne-components/issues/5207)) ([5ea4e91](https://github.com/sbb-design-systems/lyne-components/commit/5ea4e9125276d67358e85e3e8808c1cbbdf1f557)), closes [#4914](https://github.com/sbb-design-systems/lyne-components/issues/4914)
+
+
+### Bug Fixes
+
+* **core:** avoid to inert subtree of excluded elements ([#5218](https://github.com/sbb-design-systems/lyne-components/issues/5218)) ([cf366da](https://github.com/sbb-design-systems/lyne-components/commit/cf366da3ced7df36705d9c30968ceb0e2bdbb8be)), closes [#5165](https://github.com/sbb-design-systems/lyne-components/issues/5165)
+* **icon:** avoid caching failed requests ([#5208](https://github.com/sbb-design-systems/lyne-components/issues/5208)) ([ce8acb8](https://github.com/sbb-design-systems/lyne-components/commit/ce8acb81fc287dd94a41894b4da905975542064c)), closes [#5187](https://github.com/sbb-design-systems/lyne-components/issues/5187)
+
+
+### Documentation
+
+* add i18n guide ([#5210](https://github.com/sbb-design-systems/lyne-components/issues/5210)) ([8e28cd7](https://github.com/sbb-design-systems/lyne-components/commit/8e28cd7ac03b993ec9f0dc6dc3599e9c5798e05f))
+
+
+### Styles
+
+* **calendar:** allow flexible width when amount is one ([#5216](https://github.com/sbb-design-systems/lyne-components/issues/5216)) ([4f5151b](https://github.com/sbb-design-systems/lyne-components/commit/4f5151b334f11912c90ee89b2ebd79759b14385b)), closes [#5214](https://github.com/sbb-design-systems/lyne-components/issues/5214)
+* remove SBB font from off-brand theme and provide theme customization API ([#5204](https://github.com/sbb-design-systems/lyne-components/issues/5204)) ([9c146da](https://github.com/sbb-design-systems/lyne-components/commit/9c146da5ed4967967c7e657b7daf42a794dfd68f))
+
 ## [5.7.0](https://github.com/sbb-design-systems/lyne-components/compare/v5.6.0...v5.7.0) (2026-08-31)